requirements.txt
alembic==1.4.3
appdirs==1.4.4
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
Mako==1.1.3
MarkupSafe==1.1.1
oauth2client==4.1.3
Pillow==8.0.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
pypiwin32==223
Pyrebase4==4.4.3
python-dateutil==1.5
python-editor==1.0.4
python-jwt==2.0.1
pyttsx3==2.90
pytz==2020.4
pywin32==300
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
virtualenv==20.1.0
Werkzeug==1.0.1
WTForms==2.3.3
